+++ Intern/rayx-core/src/Beamline/LightSource.cpp
@@ -17,17 +17,17 @@ LightSource::LightSource(const DesignObject& dobj) {
 
 Misalignment LightSource::getMisalignmentParams() const { return m_misalignmentParams; }
 
-[[maybe_unused]] double LightSource::getPhotonEnergy() const { return m_EnergyDistribution.getAverage(); }
+[[maybe_unused]] float LightSource::getPhotonEnergy() const { return m_EnergyDistribution.getAverage(); }
 
 // needed for many of the light sources, from two angles to one direction vector
-glm::dvec3 LightSource::getDirectionFromAngles(const double phi, const double psi) const {
-    double al = cos(psi) * sin(phi);
-    double am = -sin(psi);
-    double an = cos(psi) * cos(phi);
+glm::vec3 LightSource::getDirectionFromAngles(const float phi, const float psi) const {
+    float al = cos(psi) * sin(phi);
+    float am = -sin(psi);
+    float an = cos(psi) * cos(phi);
     return {al, am, an};
 }
 
 //  (see RAYX.FOR select_energy)
-double LightSource::selectEnergy() const { return m_EnergyDistribution.selectEnergy(); }
+float LightSource::selectEnergy() const { return m_EnergyDistribution.selectEnergy(); }
 
 }  // namespace RAYX
